ml_service/requirements.txt
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
nltk==3.8.1
spacy==3.6.1
transformers==4.33.2
torch==2.0.1
sentence-transformers==2.2.2
flask==2.3.3
python-dotenv==1.0.0
pymongo==4.5.0
